+++ forge/src/main/java/net/mehvahdjukaar/every_compat/forge/EveryCompatForge.java
@@ -4,14 +4,12 @@ import net.mehvahdjukaar.every_compat.EveryCompat;
 import net.mehvahdjukaar.every_compat.EveryCompatClient;
 import net.mehvahdjukaar.every_compat.api.CompatModule;
 
-import net.mehvahdjukaar.every_compat.modules.another_furniture.AnotherFurnitureModule;
-import net.mehvahdjukaar.every_compat.modules.camp_chair.CampChairModule;
-import net.mehvahdjukaar.every_compat.modules.decorative_blocks.DecorativeBlocksModule;
-import net.mehvahdjukaar.every_compat.modules.exline.BarkCarpetsModule;
 import net.mehvahdjukaar.every_compat.modules.forge.abnormal.WoodworksModule;
 import net.mehvahdjukaar.every_compat.modules.forge.architect_palette.ArchitectsPaletteModule;
 import net.mehvahdjukaar.every_compat.modules.forge.backpacked.BackpackedModule;
+import net.mehvahdjukaar.every_compat.modules.forge.builders_delight.BuildersDelightModule;
 import net.mehvahdjukaar.every_compat.modules.forge.buildersaddition.BuildersAdditionModule;
+import net.mehvahdjukaar.every_compat.modules.forge.corail.WoodcutterModule;
 import net.mehvahdjukaar.every_compat.modules.forge.create.CreateModule;
 import net.mehvahdjukaar.every_compat.modules.forge.dramaticdoors.DramaticDoorsMacawModule;
 import net.mehvahdjukaar.every_compat.modules.forge.dramaticdoors.DramaticDoorsModule;
@@ -33,8 +31,6 @@ import net.mehvahdjukaar.every_compat.modules.forge.valhelsia.ValhelsiaStructure
 import net.mehvahdjukaar.every_compat.modules.forge.woodster.WoodsterModule;
 import net.mehvahdjukaar.every_compat.modules.forge.workshop.WorkshopForHandsomeAdventurerModule;
 import net.mehvahdjukaar.every_compat.modules.forge.xerca.XercaModule;
-import net.mehvahdjukaar.every_compat.modules.forge.builders_delight.BuildersDelightModule;
-
 import net.mehvahdjukaar.moonlight.api.platform.PlatformHelper;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
@@ -63,7 +59,7 @@ public class EveryCompatForge extends EveryCompat {
 
         CraftingHelper.register(new BlockTypeEnabledCondition.Serializer());
 
-
+        // ========================================= Add Modules ==================================================== \\
         addModule("backpacked", () -> BackpackedModule::new);
         addModule("buildersaddition", () -> BuildersAdditionModule::new);
         addModule("cfm", () -> MrCrayfishFurnitureModule::new);
@@ -87,6 +83,13 @@ public class EveryCompatForge extends EveryCompat {
         addModule("valhelsia_structures", () -> ValhelsiaStructuresModule::new);
         addModule("woodster", () -> WoodsterModule::new);
         addModule("woodworks", () -> WoodworksModule::new);
+        addModule("architects_palette", () -> ArchitectsPaletteModule::new);
+        addModule("workshop_for_handsome_adventurer", () -> WorkshopForHandsomeAdventurerModule::new);
+        addModule("xercamod", () -> XercaModule::new);
+        addModule("exnihilosequentia", () -> ExNihiloSequentiaModule::new);
+        addModule("buildersdelight", () -> BuildersDelightModule::new);
+
+        // Macaw's
         addModule("mcwbridges", () -> MacawBridgesModule::new);
         addModule("mcwdoors", () -> MacawDoorsModule::new);
         addModule("mcwfences", () -> MacawFencesModule::new);
@@ -96,19 +99,18 @@ public class EveryCompatForge extends EveryCompat {
         addModule("mcwroofs", () -> MacawRoofsModule::new);
         addModule("mcwtrpdoors", () -> MacawTrapdoorsModule::new);
         addModule("mcwwindows", () -> MacawWindowsModule::new);
-        addModule("architects_palette", () -> ArchitectsPaletteModule::new);
-		addModule("workshop_for_handsome_adventurer", () -> WorkshopForHandsomeAdventurerModule::new);
-        addModule("xercamod", () -> XercaModule::new);
-        addModule("exnihilosequentia", () -> ExNihiloSequentiaModule::new);
 
-        addModule("buildersdelight", () -> BuildersDelightModule::new);
+        // ========================================== WORK IN PROGRESS ============================================== \\
+//        addModule("carpenter", () -> CarpenterModule::new);
+//        addModule("corail_woodcutter", () -> WoodcutterModule::new);
 
+        // ======================================== DISABLED FOR A REASON =========================================== \\
 
-        // Disabled due to block entity complications
-
-        // Disabled until custom block models work
+        // =============================== DISABLED UNTIL CUSTOM BLOCK MODELS WORK ================================== \\
             // addModule("graveyard", () -> GraveyardModule::new);
 
+
+        // ========================================================================================================== \\
         if (PlatformHelper.getEnv().isClient()) {
             EveryCompatClient.commonInit();
         }