+++ changelog.md
@@ -1,4 +1,2 @@
 #### UPDATED:
-- Macaw's Windows Module to support **v2.2.1** 
----
-**NOTE:** Older than v2.2.1 of Macaw's Windows is not supported anymore
+- XercaMod - Fixed Recipes & Improved Textures